app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
automat==22.10.0
    # via twisted
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r requirements.in
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
constantly==15.1.0
    # via twisted
cryptography==39.0.1
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.2.0
    # via
    #   parsel
    #   scrapy
django==4.1.6
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
django-decouple==2.1
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
filelock==3.9.0
    # via tldextract
gunicorn==20.1.0
    # via -r requirements.in
hyperlink==21.0.0
    # via twisted
idna==3.4
    # via
    #   hyperlink
    #   requests
    #   tldextract
incremental==22.10.0
    # via twisted
itemadapter==0.7.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.0.6
    # via scrapy
jmespath==1.0.1
    # via itemloaders
kombu==5.2.4
    # via celery
lxml==4.9.2
    # via
    #   parsel
    #   scrapy
packaging==23.0
    # via
    #   parsel
    #   scrapy
parsel==1.7.0
    # via
    #   itemloaders
    #   scrapy
pika==1.3.1
    # via -r requirements.in
pillow==9.4.0
    # via -r requirements.in
prompt-toolkit==3.0.36
    # via click-repl
protego==0.2.1
    # via scrapy
psycopg2-binary==2.9.5
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.21
    # via cffi
pydispatcher==2.0.6
    # via scrapy
pyopenssl==23.0.0
    # via scrapy
pytz==2022.7.1
    # via
    #   celery
    #   djangorestframework
queuelib==1.6.2
    # via scrapy
redis==4.5.1
    # via -r requirements.in
requests==2.28.2
    # via
    #   requests-file
    #   tldextract
requests-file==1.5.1
    # via tldextract
scrapy==2.8.0
    # via -r requirements.in
service-identity==21.1.0
    # via scrapy
six==1.16.0
    # via
    #   automat
    #   click-repl
    #   protego
    #   requests-file
    #   service-identity
sqlparse==0.4.3
    # via django
tldextract==3.4.0
    # via scrapy
twisted==22.10.0
    # via scrapy
typing-extensions==4.4.0
    # via twisted
urllib3==1.26.14
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
w3lib==2.1.1
    # via
    #   itemloaders
    #   parsel
    #   scrapy
wcwidth==0.2.6
    # via prompt-toolkit
zope-interface==5.5.2
    # via
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
